requirements.txt
# Core Dash dependencies
dash>=2.14.0
dash-bootstrap-components>=1.5.0
dash-leaflet>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Geospatial
geopy>=2.4.0

# Visualization
plotly>=5.18.0

# Optional: for better performance
dash-extensions>=1.0.0

# Firebase Authentication & Database
pyrebase4>=4.8.0
firebase-admin>=7.0.0
